refactor: modularize graph panel into sub-components for labels, thresholds, autogrid, bounds, and overlays
