+++ src/ui/panels/graph/bounds.rs
@@ -1,0 +1,150 @@
+/*
+ * Copyright 2025 Federico D'Ambrosio
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+use crate::app::{PanelState, YAxisMode};
+
+pub(super) fn calculate_y_bounds(p: &PanelState) -> [f64; 2] {
+    let mut min = f64::MAX;
+    let mut max = f64::MIN;
+    let mut has_data = false;
+
+    for s in &p.series {
+        if !s.visible {
+            continue;
+        }
+        for &(_, v) in &s.points {
+            if !v.is_finite() {
+                continue;
+            }
+            if v < min {
+                min = v;
+            }
+            if v > max {
+                max = v;
+            }
+            has_data = true;
+        }
+    }
+
+    if !has_data {
+        return [0.0, 1.0];
+    }
+
+    if min == max {
+        min -= 1.0;
+        max += 1.0;
+    }
+
+    if p.y_axis_mode == YAxisMode::ZeroBased {
+        if min > 0.0 {
+            min = 0.0;
+        } else if max < 0.0 {
+            max = 0.0;
+        }
+    }
+
+    // Add some padding
+    let range = max - min;
+    [min - range * 0.05, max + range * 0.05]
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::app::SeriesView;
+
+    fn create_test_panel() -> PanelState {
+        PanelState {
+            title: "test".to_string(),
+            exprs: vec![],
+            legends: vec![],
+            series: vec![],
+            last_error: None,
+            last_url: None,
+            last_samples: 0,
+            grid: None,
+            y_axis_mode: YAxisMode::Auto,
+            panel_type: crate::app::PanelType::Graph,
+            thresholds: None,
+            min: None,
+            max: None,
+            autogrid: None,
+        }
+    }
+
+    #[test]
+    fn test_calculate_y_bounds_basic() {
+        let mut p = create_test_panel();
+        p.series.push(SeriesView {
+            name: "test".to_string(),
+            value: None,
+            points: vec![(0.0, 10.0), (1.0, 20.0)],
+            visible: true,
+        });
+
+        let bounds = calculate_y_bounds(&p);
+        assert!(bounds[0] < 10.0);
+        assert!(bounds[1] > 20.0);
+    }
+
+    #[test]
+    fn test_calculate_y_bounds_nan() {
+        let mut p = create_test_panel();
+        p.series.push(SeriesView {
+            name: "test".to_string(),
+            value: None,
+            points: vec![(0.0, 10.0), (1.0, f64::NAN), (2.0, 20.0)],
+            visible: true,
+        });
+
+        let bounds = calculate_y_bounds(&p);
+        assert!(bounds[0] < 10.0); // Should ignore NAN
+        assert!(bounds[1] > 20.0);
+    }
+
+    #[test]
+    fn test_calculate_y_bounds_infinity() {
+        let mut p = create_test_panel();
+        p.series.push(SeriesView {
+            name: "test".to_string(),
+            value: None,
+            points: vec![(0.0, 10.0), (1.0, f64::INFINITY), (2.0, 20.0)],
+            visible: true,
+        });
+
+        let bounds = calculate_y_bounds(&p);
+        assert!(bounds[0] < 10.0); // Should ignore INFINITY
+        assert!(bounds[1] > 20.0);
+    }
+
+    #[test]
+    fn test_calculate_y_bounds_zero_based() {
+        let mut p = create_test_panel();
+        p.y_axis_mode = YAxisMode::ZeroBased;
+        p.series.push(SeriesView {
+            name: "test".to_string(),
+            value: None,
+            points: vec![(0.0, 10.0), (1.0, 20.0)],
+            visible: true,
+        });
+
+        let bounds = calculate_y_bounds(&p);
+        // Range is 0.0 to 20.0. Padding is 5% of 20.0 = 1.0.
+        // So min should be 0.0 - 1.0 = -1.0.
+        assert_eq!(bounds[0], -1.0);
+        assert!(bounds[1] > 20.0);
+    }
+}
